fix email heuristic
