+++ guarddog/analyzer/metadata/potentially_compromised_email_domain.py
@@ -2,6 +2,7 @@
 
 Detects if a maintainer's email domain might have been compromised.
 """
+from abc import abstractmethod
 from datetime import datetime
 from typing import Optional
 
@@ -12,6 +13,10 @@ from guarddog.analyzer.metadata.detector import Detector
 
 class PotentiallyCompromisedEmailDomainDetector(Detector):
     RULE_NAME = "potentially_compromised_email_domain"
+
+    def __init__(self):
+        super().__init__()
+        self.ecosystem = ""
 
     def _get_domain_creation_date(self, email_domain) -> tuple[Optional[datetime], bool]:
         """
@@ -45,3 +50,57 @@ class PotentiallyCompromisedEmailDomainDetector(Detector):
             return min(creation_dates), True
 
         return creation_dates, True
+
+
+    def detect(self, package_info, path: Optional[str] = None) -> tuple[bool, str]:
+        """
+        Uses a package's information from PyPI's JSON API to determine
+        if the package's email domain might have been compromised
+
+        Args:
+            package_info (dict): dictionary representation of PyPI's JSON
+                output
+
+        Raises:
+            Exception: "Email for {package_info['info']['name']} does not exist."
+
+        Returns:
+            bool: True if email domain is compromised
+        """
+
+        emails = self.get_email_addresses(package_info)
+
+        if len(emails) == 0:
+            # No e-mail is set for this package, hence no risk
+            return False, "No e-mail found for this package"
+
+        latest_project_release = self.get_project_latest_release_date(package_info)
+
+        has_issues = False
+        messages = []
+        for email in emails:
+            sanitized_email = email.strip().replace(">", "").replace("<", "")
+            email_domain = sanitized_email.split("@")[-1]
+            domain_creation_date, domain_exists = self._get_domain_creation_date(email_domain)
+
+            if not domain_exists:
+                has_issues = True
+                messages.append(f"The maintainer's email ({email}) domain does not exist and can likely be registered "
+                                f"by an attacker to compromise the maintainer's {self.ecosystem} account")
+            if domain_creation_date is None or latest_project_release is None:
+                continue
+            if latest_project_release < domain_creation_date:
+                has_issues = True
+                messages.append(f"The domain name of the maintainer's email address ({email}) was"" re-registered after"
+                                " the latest release of this ""package. This can be an indicator that this is a"""
+                                " custom domain that expired, and was leveraged by"" an attacker to compromise the"
+                                f" package owner's {self.ecosystem}"" account.")
+        return has_issues, "\n".join(messages)
+
+    @abstractmethod
+    def get_project_latest_release_date(self, package_info):
+        pass
+
+    @abstractmethod
+    def get_email_addresses(self, package_info):
+        pass
